requirements.txt
cowsay
requests
pillow
pyTelegramBotAPI